\name{adaptsmoFMRI-package}
\alias{adaptsmoFMRI-package}
\alias{adaptsmoFMRI}
\docType{package}
\title{Adaptive Smoothing of FMRI Data}
\description{
This package contains R functions for estimating the blood
oxygenation level dependent (BOLD) effect by using functional 
Magnetic Resonance Imaging (fMRI) data, based on 
adaptive Gauss Markov random fields, for real as well as 
simulated data. The implemented simulations make use of efficient 
Markov Chain Monte Carlo methods. 
}
\references{
A. Brezger, L. Fahrmeir, A. Hennerfeind
  \emph{Adaptive Gaussian Markov random fields with applications in
    human brain mapping}, Applied Statistics, \bold{56}, Part 3,
  pp. 327-345, 2007

V. J. Schmid, A. Thornton, B. Witcher 
  \emph{Working with the DICOM and NIfTI Data Standards in R}, 
Journal of Statistical Software, \bold{Volume 44}, Issue 6. 2011
}
\details{
\tabular{ll}{
Package: \tab adaptsmoFMRI\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-01-16\cr
License: \tab GPL-2\cr
}
}
\author{
Max Hughes
Maintainer: \email{hughesgm@me.com}
}
\keyword{ package }
