\name{data_simfmri}
\docType{data}
\alias{data_simfmri}
\title{Simulated FMRI Data Set (for one covariate)}
\description{The data set contains a synthetic spatiotemporal data set, hence a 
simulated data set. The data set is an array of dimension 20 x 20 x 210.}
\note{This data set was created while inserting one covariate.}
