\name{data_simfmri2COVAR}
\docType{data}
\alias{data_simfmri2COVAR}
\title{Simulated FMRI Data Set (for two covariates)}
\description{The data set contains a synthetic spatiotemporal data set, hence a 
simulated data set. The data set is an array of dimension 20 x 20 x 180.}
\note{This data set was created while inserting two covariates.}